ai_service/requirements.txt
# Reminisce AI Brain Dependencies
# Install with: pip install -r requirements.txt

# Pinecone vector database client
pinecone-client>=3.0.0

# LangChain with Google Vertex AI integration
langchain-google-vertexai>=1.0.0

# Google Cloud AI Platform (for embeddings)
google-cloud-aiplatform>=1.38.0

# Vertex AI SDK
vertexai>=1.0.0

